css files moved into components folders
