--- src/CSSForComponents/AppSidebar.css
@@ -1,1 +1,0 @@
-
