+++ src/ComponentsDashboard/LoginPages/Input.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import '../../CSSForComponents/LoginPage.css';
-import '../../CSSForComponents/SignUpPage.css';
+import './LoginPage.css';
+import './SignUpPage.css';
 
 export const Input = ({ type, placeholder }) => {
   return <input className="input" type={type} placeholder={placeholder}></input>;
