+++ src/ComponentsDashboard/LoginPages/SignInPage.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import '../../CSSForComponents/App.css';
-import '../../CSSForComponents/LoginPage.css';
+import '../../App.css';
+import './LoginPage.css';
 import { PopeyeImg } from '../Dashboards/PopeyeImg';
 import { LoginPageFooter } from './LoginFooter';
 import { Input } from './Input';
 import LoginButton from './LoginButton';
 import { NameOfApp } from '../Dashboards/NameOfApp';
 import { Divider } from '@material-ui/core';
-import '../../CSSForComponents/SignUpPage.css';
+import './SignUpPage.css';
 // import { OrLine } from "./Orline";
 
 const SignInPage = () => {
